requirements.txt

bc-jsonpath-ng==1.6.1
bc-python-hcl2==0.4.2
beartype==0.19.0
beautifulsoup4==4.12.3
boolean.py==4.0
boto3==1.35.49
botocore==1.35.58
build==1.2.2.post1
CacheControl==0.14.2
cached-property==2.0.1
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
cfgv==3.4.0
charset-normalizer==3.4.0
checkov==3.2.289
click==8.1.7
click-option-group==0.5.6
cloudsplaining==0.7.0
colorama==0.4.6
ConfigArgParse==1.7
contextlib2==21.6.0
cryptography==43.0.3
cyclonedx-python-lib==6.4.4
dacite==1.8.1
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.15
dill==0.3.6
distlib==0.3.9
dnslib==0.9.25
dnspython==2.7.0
docker==7.1.0
dockerfile-parse==2.0.1
dpath==2.1.3
ecdsa==0.19.0
exceptiongroup==1.2.2
fastapi==0.103.2
filelock==3.16.1
firebase-admin==6.6.0
frozenlist==1.5.0

gitdb==4.0.11
GitPython==3.1.43
google-api-core==2.24.1
google-api-python-client==2.162.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.2
google-cloud-firestore==2.20.1
google-cloud-storage==3.1.0
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.68.0
grpcio==1.70.0
grpcio-status==1.70.0
h11==0.14.0
httplib2==0.22.0
identify==2.6.2
idna==3.10
importlib_metadata==7.2.1
isodate==0.7.2
jaconv==0.4.0
Jinja2==3.1.4
jmespath==1.0.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
junit-xml==1.9
lark==1.2.2
license-expression==30.4.0
localstack==3.8.1
localstack-core==3.8.1
localstack-ext==3.8.1

Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
msgpack==1.1.0
multidict==6.1.0
mutagen==1.47.0
networkx==2.6.3
nodeenv==1.9.1
numpy==2.0.2
openai==0.28.1
orjson==3.10.11
packageurl-python==0.13.4
packaging==23.2
platformdirs==4.3.6
plux==1.12.1
ply==3.11
policy_sentry==0.13.1
pre_commit==4.0.1
prettytable==3.12.0
propcache==0.2.0
proto-plus==1.26.0
protobuf==5.29.3
psutil==6.1.0
py-serializable==1.1.2
pyaes==1.6.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycares==4.4.0
pycep-parser==0.5.1
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
PyJWT==2.10.1
pykakasi==2.3.0
pyotp==2.9.0
pyparsing==3.2.0
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
python-slugify==8.0.4
pytube==15.0.0
PyYAML==6.0.2
RapidFuzz==3.10.1
rdflib==7.1.1
redis==5.2.0
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
rich==13.9.4
rpds-py==0.21.0
rsa==4.9
rustworkx==0.13.2
s3transfer==0.10.3
schema==0.7.5
semantic-version==2.10.0
semver==3.0.2

smmap==5.0.1
sniffio==1.3.1
sortedcontainers==2.4.0
soundcloud-v2==1.6.0
soupsieve==2.6
spdx-tools==0.8.3
spotdl==4.2.9
spotipy==2.24.0
starlette==0.27.0
syncedlyrics==1.0.1
tabulate==0.9.0
tailer==0.4.1
termcolor==2.3.0
text-unidecode==1.3
tomli==2.1.0
tqdm==4.67.0
typing_extensions==4.12.2
Unidecode==1.3.8
unidiff==0.7.5
uritemplate==4.1.1
uritools==4.0.3
urllib3==1.26.20
uvicorn==0.23.2
virtualenv==20.27.1
wcwidth==0.2.13
wrapt==1.16.0
xmltodict==0.14.2
yarl==1.17.1
yt-dlp==2024.11.18
ytmusicapi==1.8.2
zipp==3.21.0
